Close #1430 by applying warnings fix patch.
git-svn-id: http://svn.mythtv.org/svn/trunk@10027 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythtranscode/mpeg2fix.cpp
@@ -1859,7 +1859,7 @@ int MPEG2fixup::Start()
                                     PTSdiscrep = 0;
                                     break;
                                 }
-                                if (tmpPTSdiscrep != AV_NOPTS_VALUE &&
+                                if (tmpPTSdiscrep != (int64_t)AV_NOPTS_VALUE &&
                                     tmpPTSdiscrep != PTSdiscrep)
                                     PTSdiscrep = tmpPTSdiscrep;
                             }
